Added Validators and ValidationChain components
Added a few helpers for validating use input in gwt. The Validator
chain will allow to chain multiple validators together. Had to make
use of generics in order to have validators implement a single
interface.

Had to make adjustments to the AdvancedTextbox as well to ensure
that the value returned using getText() returns an empty string
if the placeholder text is still present.

issue #158