Added DynamicForm and fixed FornValidator
Added the DynamicForm and added the TextInputField that can be used with the
DynamicForm.  The DynamicForm still needs a function to set the InputField to
readonly mode.  Also fixed the FormValidator.  Also added a README for the
DynamicForm that shows an example of how to use the DynamicForm will be change
explain the different fields that work with the DynamicForm later.

issue #49