Added SSDateBox and SSDatePicker
Added the SSDateBox and SSDatePicker that extend the default GWT DateBox and
DatePicker. But the way the GWT DateBox and DatePicker is written we are force
to include Google's DateBox and DatePicker code in our datapicker package. All
code copied from Google remains with the Google license in the files.

issue #49