Fixed an issue on the LeftMenu on IE
Fixed an issue on the LeftMenu on IE that only happened on IE with one menu item
It seems like the issue only happens IE the animate is called before the menu is
rendered for the first time. The deferred schedule command is calls the animate
at a later stage after the menu has been rendered.

issue A24Group/Triage#437