Fixed the bug on actions in the shift entries/expenses action column
The list of actions got overriden each time the cell is loaded. This means
that any action that is not in the last cell would never fire a mouse up
event. Created a new array list which will be used for display purpose only
while the old action list will be used to determine what was clicked.

Issue A24Group/Triage#4021